inference/pyproject.toml
[project]
name = "dstk-46-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "catboost>=1.2.10",
    "fastapi>=0.135.3",
    "google-cloud-storage>=3.10.1",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
]
